requirements.txt
# Database
psycopg2-binary
SQLAlchemy
python-dotenv

# Data Processing
pandas
numpy

# Utilities
tqdm  # Progress bars